requirements.txt
pandas==2.2.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyyaml==6.0.1
websockets==10.4        
base58==2.1.1
solders
solana
anchorpy
construct
asyncio
